Created tools directory under quads and moved tools there
